pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orange3-imagenets"
version = "0.0.2"
description = "A collection of Orange3 widgets for training and applying neural networks to images"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "CC-BY-NC-SA-4.0" }
authors = [
    { name = "Chris Lee", email = "github@chrislee.dhs.org" },
]
keywords = ["orange3 add-on"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3 :: Only"
]
dependencies = [
    "Orange3 >= 3.39.0",
    "orange-canvas-core >= 0.1.28",
    "orange-widget-base >= 4.20.0",
    "Orange3-ImageAnalytics >= 0.13.0",
    "opencv-python-headless >= 4.12",
    "keras >= 3.10.0",
    "tf_keras >= 2.19.0"
]

[license]
license = "CC-BY-NC-SA-4.0"
license-files = ["LICENSE"]

[project.optional-dependencies]
test = ["coverage"]
doc = ["sphinx", "recommonmark", "sphinx_rtd_theme"]

[project.urls]
Homepage = "https://github.com/chrislee35/orange3-imagenets"

[tool.setuptools.packages.find]
where = ["."]
namespaces = true

[tool.setuptools.package-data]
"orangecontrib.imagenets.widgets" = ["icons/*.svg"]

[tool.setuptools.data-files]
"help/orange3-imagenets" = ["doc/_build/html/**/*"]

[project.entry-points."orange.widgets"]
imagenets = "orangecontrib.imagenets.widgets"

[project.entry-points."orange3.addon"]
orange3-imagenets = "orangecontrib.imagenets"

[project.entry-points."orange.canvas.help"]
html-index = "orangecontrib.imagenets.widgets:WIDGET_HELP_PATH"
